SDK-537: Move out profile tests
